+++ changelog.md
@@ -64,6 +64,7 @@
 - Support TLS encrypted connections to PostgreSQL when using
   `vmq_diversity`. Fixes #811.
 - Support TLS encrypted connections to MongoDB when using `vmq_diversity`.
+- Add CockroachDB support to the `vmq_diversity` plugin.
 
 ## VerneMQ 1.7.0